refactor(atelier_vapor): split lib into cohesive modules
